refactoring rules into derived classes
